Add test and move it and the mapillary one to the new package locations
