--- src/androidTest/java/de/blau/android/layer/mapillary/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Mapillary related testing
- */
-package de.blau.android.layer.mapillary;
