+++ src/androidTest/java/de/blau/android/layer/streetlevel/mapillary/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Mapillary related testing
+ */
+package de.blau.android.layer.streetlevel.mapillary;
