+++ src/androidTest/java/de/blau/android/layer/streetlevel/panoramax/package-info.java
@@ -1,0 +1,1 @@
+package de.blau.android.layer.streetlevel.panoramax;
