+++ src/testCommon/resources/fixtures/panoramax_sequences.yaml
@@ -1,0 +1,4 @@
+statusCode: 200       
+delay: 0              
+body: 'panoramax_sequences.json'
+
